Skill section moved to the appropriate folder
