+++ src/components/skill/backendSkill/BackendSkill.js
@@ -1,11 +1,11 @@
-import Skill from "../../skill/Skill";
+import Skill from "../Skill";
 import Badge from "../../badge/Badge";
 import nodeBadge from "../../../assets/badges/node.svg";
 import expressBadge from "../../../assets/badges/express.svg";
 import csharpBadge from "../../../assets/badges/csharp.svg";
 import firebaseBadge from "../../../assets/badges/firebase.svg";
 import firestoreBadge from "../../../assets/badges/firestore.svg";
-import "./BackendSkills.scss";
+import "./BackendSkill.scss";
 
 const BackendSkill = () => {
   const nodeBadgeElement = Badge({ src: nodeBadge });
