+++ src/components/skill/uxSkill/UxSkill.scss
